# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wood Trading API",
        "description": "API for Wood Trading Application - A comprehensive system for managing wood inventory, sales, purchases, and financial operations.",
        "contact": {
            "name": "Wood Trading Support",
            "email": "support@woodtrading.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.woodtrading.com/api",
            "description": "Production Server"
        },
        {
            "url": "D:\\Haleem-dev\\wood-trading"
        }
    ],
    "paths": {
...
        "/api/stock": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get stock levels",
                "operationId": "1a8e72c4f6a7a5f6ec0996d1be80fa3d",
                "parameters": [
                    {
                        "name": "item_id",
                        "in": "query",
                        "description": "Filter by item ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Filter by warehouse ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "low_stock",
                        "in": "query",
                        "description": "Show only low stock items",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stock levels retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/StockLotResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/stock/adjust": {
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Adjust stock levels",
                "operationId": "1e9979969dcd2cb5e948dde1d1c28921",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item_id",
                                    "warehouse_id",
                                    "quantity",
                                    "type",
                                    "reason"
                                ],
                                "properties": {
                                    "item_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.5
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "adjustment"
                                        ],
                                        "example": "adjustment"
                                    },
                                    "movement": {
                                        "type": "string",
                                        "enum": [
                                            "in",
                                            "out"
                                        ],
                                        "example": "in"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Physical count adjustment"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Found additional stock during inventory"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock adjusted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Stock adjusted successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/StockLotResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StockLotResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "lot_no": {
                        "type": "string",
                        "example": "LOT2024001"
                    },
                    "qty_on_hand": {
                        "type": "integer",
                        "example": 100
                    },
                    "qty_reserved": {
                        "type": "integer",
                        "example": 10
                    },
                    "qty_available": {
                        "type": "integer",
                        "example": 90
                    },
                    "cost": {
                        "type": "number",
                        "format": "float",
                        "example": 25.5
                    },
                    "received_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "warehouse": {
                        "$ref": "#/components/schemas/WarehouseResource"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "WarehouseResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Main Warehouse"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Industrial Blvd, City, State"
                    },
                    "contact_person": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1-555-0456"
                    },
                    "email": {
                        "type": "string",
                        "example": "warehouse@company.com"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in the format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}